+++ iOS/G3MiOSSDK/Specific/Derived/Canvas_iOS.mm
@@ -331,7 +331,8 @@ const Vector2F Canvas_iOS::_textExtent(const std::string& text) {
 
   CGSize cgSize = [nsString sizeWithFont: _currentUIFont];
 
-  return Vector2F(cgSize.width, cgSize.height);
+  return Vector2F((float) cgSize.width,
+                  (float) cgSize.height);
 }
 
 void Canvas_iOS::_fillText(const std::string& text,
